refactor filenames in types directory
