+++ project/src/index.tsx
@@ -9,7 +9,7 @@ import {reducer} from './store/reducer';
 import {composeWithDevTools} from 'redux-devtools-extension';
 import {requireAuthorization} from './store/actions';
 import {AuthorizationStatus} from './const';
-import {ThunkAppDispatch} from './types/action-types';
+import {ThunkAppDispatch} from './types/action';
 import {checkAuthAction, fetchHotelsAction} from './store/api-actions';
 
 const MainSetting = {
