+++ project/src/store/api-actions.ts
@@ -1,10 +1,10 @@
-import {ThunkActionResult} from '../types/action-types';
+import {ThunkActionResult} from '../types/action';
 import {loadOffers, requireAuthorization, requireLogout} from './actions';
 import {saveToken, dropToken} from '../services/token';
 import {APIRoute, AuthorizationStatus} from '../const';
 import {Offers} from '../types/types';
 import {AuthData} from '../types/auth-data';
-import {Token} from '../types/api-types';
+import {Token} from '../types/api';
 import {adaptOffersToClient} from '../utils';
 
 export const fetchHotelsAction = (): ThunkActionResult =>
